+++ src/main/java/no/anderska/wta/QuestionChecker.java
@@ -12,5 +12,5 @@ public interface QuestionChecker {
 
     List<QuestionCategory> allCategories();
 
-    List<Question> listCategory(long categoryId);
+    List<Question> listCategory(int categoryId);
 }
